[*]修改一部分逻辑 添加游戏流程

+++ src/Practice/CozyAdventure.Protocol/Msg/FarmIncomeMessage.cs
@@ -12,9 +12,9 @@ namespace CozyAdventure.Protocol.Msg
     {
         public override uint Id { get { return (uint)MessageId.Farm.FarmIncomeMessage; } }
 
-        public int Money { get; set; }
+        public long Money { get; set; }
 
-        public int Exp { get; set; }
+        public long Exp { get; set; }
 
         public int Item { get; set; }
     }
